[FLINK-36082][pipeline-connector][kafka] Fix lamda NotSerializableException in KafkaDataSink
This closes #3549.
