+++ flink-cdc-connect/flink-cdc-pipeline-connectors/flink-cdc-pipeline-connector-kafka/src/main/java/org/apache/flink/cdc/connectors/kafka/sink/KafkaDataSink.java
@@ -19,9 +19,6 @@ package org.apache.flink.cdc.connectors.kafka.sink;
 
 import org.apache.flink.api.common.serialization.SerializationSchema;
 import org.apache.flink.cdc.common.event.Event;
-import org.apache.flink.cdc.common.event.SchemaChangeEvent;
-import org.apache.flink.cdc.common.event.SchemaChangeEventType;
-import org.apache.flink.cdc.common.event.SchemaChangeEventTypeFamily;
 import org.apache.flink.cdc.common.sink.DataSink;
 import org.apache.flink.cdc.common.sink.EventSinkProvider;
 import org.apache.flink.cdc.common.sink.FlinkSinkProvider;
@@ -33,10 +30,7 @@ import org.apache.flink.connector.kafka.sink.KafkaSinkBuilder;
 import org.apache.kafka.clients.producer.ProducerConfig;
 
 import java.time.ZoneId;
-import java.util.Arrays;
 import java.util.Properties;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 /** A {@link DataSink} for "Kafka" connector. */
 public class KafkaDataSink implements DataSink {
@@ -104,33 +98,6 @@ public class KafkaDataSink implements DataSink {
 
     @Override
     public MetadataApplier getMetadataApplier() {
-        return new MetadataApplier() {
-
-            private Set<SchemaChangeEventType> enabledEventTypes =
-                    Arrays.stream(SchemaChangeEventTypeFamily.ALL).collect(Collectors.toSet());
-
-            @Override
-            public MetadataApplier setAcceptedSchemaEvolutionTypes(
-                    Set<SchemaChangeEventType> schemaEvolutionTypes) {
-                enabledEventTypes = schemaEvolutionTypes;
-                return this;
-            }
-
-            @Override
-            public boolean acceptsSchemaEvolutionType(SchemaChangeEventType schemaChangeEventType) {
-                return enabledEventTypes.contains(schemaChangeEventType);
-            }
-
-            @Override
-            public Set<SchemaChangeEventType> getSupportedSchemaEvolutionTypes() {
-                // All schema change events are supported.
-                return Arrays.stream(SchemaChangeEventTypeFamily.ALL).collect(Collectors.toSet());
-            }
-
-            @Override
-            public void applySchemaChange(SchemaChangeEvent schemaChangeEvent) {
-                // simply do nothing here because Kafka do not maintain the schemas.
-            }
-        };
+        return new KafkaMetaDataApplier();
     }
 }
